+++ CHANGELOG.md
@@ -8,6 +8,20 @@ and this project "attempts" to adhere to [Semantic Versioning](http://semver.org
 
 TO DO: A true quick start example, just SRA data?
 
+## [v0.6.5] - 2020-0304
+
+### Added
+- NCBImetaUtilties.py function: adv_xml_search to allow pre-formatted XPath query (PR #9)
+- Two new error classes ErrorXPathQueryMultiElement and ErrorXPathElementUnknown
+
+### Changed
+- Bug: Improved specificty of SRABioSampleAccession and SRABioProjectAccession
+- NucleotideBioSampleAccession now a preformatted XPath query (previously empty value)
+- BioSampleBioProjectAccession now a preformatted XPath query (previously non-specific)
+
+### Removed
+- Non-specific node NucleotideAssemblyAccession
+
 ## [v0.6.4] - 2020-0220 - Bobby Tables
 
 ### Added
@@ -219,7 +233,9 @@ Jumps directly from v0.5.0 to v0.6.0 because changes are significant enough to b
 - Repository migrated from GenomeCollector
 
 [Development]: https://github.com/ktmeaton/NCBImeta/compare/HEAD...dev
-[v0.6.2]: https://github.com/ktmeaton/NCBImeta/compare/v0.6.2...HEAD
+[v0.6.4]: https://github.com/ktmeaton/NCBImeta/compare/v0.6.4...HEAD
+[v0.6.3]: https://github.com/ktmeaton/NCBImeta/compare/v0.6.3...v0.6.4
+[v0.6.2]: https://github.com/ktmeaton/NCBImeta/compare/v0.6.2...v0.6.3
 [v0.6.1]: https://github.com/ktmeaton/NCBImeta/compare/v0.6.1...v0.6.2
 [v0.6.0]: https://github.com/ktmeaton/NCBImeta/compare/v0.6.0...v0.6.1
 [v0.5.0]: https://github.com/ktmeaton/NCBImeta/compare/v0.5.0...v0.6.0